Refactor client script into modular architecture
