+++ index.html
@@ -56,7 +56,7 @@
   <link rel="stylesheet" href="style.css" />
   <script defer src="https://cdn.jsdelivr.net/npm/html2canvas@1.4.1/dist/html2canvas.min.js"></script>
   <script defer src="https://cdn.jsdelivr.net/npm/qrcode@1.5.1/build/qrcode.min.js"></script>
-  <script defer src="script.js"></script>
+  <script type="module" src="js/main.js"></script>
 </head>
 <body>
   <main class="container py-4">
