+++ js/state.js
@@ -1,0 +1,33 @@
+export const state = {
+  hardwareType: 'Screw',
+  systemType: 'Metric',
+  screwType: 'Bolt',
+  fuseType: 'Glass',
+  threadSize: '',
+  length: '',
+  fuseValue: '',
+  glassSpeed: '',
+  glassSize: '',
+  notes: '',
+  standard: '',
+  standardCode: '',
+  showStandard: true,
+  showImage: true,
+  showQr: false,
+  qrContent: '',
+  widthMm: 55,
+  heightMm: 12,
+  connectorCategory: '',
+  componentCategory: 'Resistor',
+  componentMount: 'Through-Hole',
+  bearingType: '',
+  bearingDetails: '',
+  customLine1: '',
+  customLine2: '',
+  customImageData: '',
+  customImageName: ''
+};
+
+export const standardFilterState = {
+  query: ''
+};
